[gobby-#14741] refactor: split service installer module
